requirements.txt
earthengine-api
geopandas
rasterio
rioxarray
xarray
shapely
pyproj
pydantic
click
loguru
google-cloud-storage
